refactor: 함수 구조 변경
