+++ interface-module/src/main/java/org/c4marathon/assignment/usecase/transfer/PendingTransferUseCase.java
@@ -1,12 +1,18 @@
 package org.c4marathon.assignment.usecase.transfer;
 
 import java.util.List;
+import java.util.Map;
 
+import org.c4marathon.assignment.domain.model.Member;
 import org.c4marathon.assignment.domain.model.PendingTransfer;
 import org.c4marathon.assignment.domain.model.TransferLog;
+import org.c4marathon.assignment.domain.service.MainAccountService;
 import org.c4marathon.assignment.domain.service.PendingTransferService;
+import org.c4marathon.assignment.domain.service.ReminderService;
 import org.c4marathon.assignment.domain.service.TransferLogFactory;
 import org.c4marathon.assignment.domain.service.TransferLogService;
+import org.c4marathon.assignment.infra.properties.MainAccountPolicy;
+import org.c4marathon.assignment.model.Account;
 import org.c4marathon.assignment.model.policy.ExternalAccountPolicy;
 import org.c4marathon.assignment.retry.RetryExecutor;
 import org.c4marathon.assignment.api.transfer.dto.TransferPendingRequestDto;
@@ -23,30 +29,26 @@ public class PendingTransferUseCase {
 	private final PendingTransferService pendingTransferService;
 	private final TransferLogService transferLogService;
 	private final TransferLogFactory transferLogFactory;
+	private final ReminderService reminderService;
 
-	private final AccountPolicyProperties accountPolicyProperties;
+	private final MainAccountPolicy mainAccountPolicy;
 	private final RetryExecutor retryExecutor;
 
 	@Transactional
 	public void createPendingTransfer(TransferPendingRequestDto request) {
 		Long shortfall = mainAccountService.calculateShortfall(request.fromAccountId(), request.amount());
 
-		if (shortfall <= 0) {
-			retryExecutor.executeWithRetry(
-					() -> pendingTransferService.initiate(request.fromAccountId(), request.toAccountId(),
-							request.amount()));
-			return;
+		if (shortfall > 0) {
+			long chargeAmount = mainAccountPolicy.getRoundedCharge(shortfall);
+			mainAccountService.chargeOrThrow(request.fromAccountId(), chargeAmount, request.amount());
+
+			Account toAccount = mainAccountService.getRefreshedAccount(request.fromAccountId());
+			TransferLog transferLog = transferLogFactory.createExternalChargeLog(
+				ExternalAccountPolicy.TEMPORARY_CHARGING,
+				toAccount,
+				chargeAmount);
+			transferLogService.saveTransferLog(transferLog);
 		}
-
-		long chargeAmount = mainAccountPolicy.getRoundedCharge(shortfall);
-		mainAccountService.chargeOrThrow(request.fromAccountId(), chargeAmount, request.amount());
-
-		Account toAccount = mainAccountService.getRefreshedAccount(request.fromAccountId());
-		TransferLog transferLog = transferLogFactory.createExternalChargeLog(
-			ExternalAccountPolicy.TEMPORARY_CHARGING,
-			toAccount,
-			chargeAmount);
-		transferLogService.saveTransferLog(transferLog);
 
 		PendingTransfer tx = retryExecutor.executeWithRetry(
 			() -> pendingTransferService.initiate(
@@ -117,4 +119,8 @@ public class PendingTransferUseCase {
 		}
 	}
 
+	public void remindPendingTargetTransactionsByImproved() {
+		Map<Member, List<PendingTransfer>> remindTargetGroupedByMember = pendingTransferService.findRemindTargetGroupedByMember();
+		reminderService.remindTransactions(remindTargetGroupedByMember);
+	}
 }
